requirements.txt
sentence-transformers
torch
psycopg2-binary
python-dotenv
pgvector